LPS-123427 Move saveDataDefinition function to the dataDefinition util
