+++ modules/apps/data-engine/data-engine-taglib/src/main/resources/META-INF/resources/data_layout_builder/js/data-layout-builder/DataLayoutBuilder.es.js
@@ -483,7 +483,6 @@ class DataLayoutBuilder extends React.Component {
 					]),
 				],
 			},
-			true,
 			defaultLanguageId
 		);
 		const fieldSetDataLayout = normalizeDataLayout(
