+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/util/settingsContext.es.js
@@ -98,6 +98,7 @@ export const updateSettingsContextProperty = (
 							...propertyValue,
 						};
 					}
+
 					field.localizedValue = {
 						...field.localizedValue,
 						[editingLanguageId]: value,
